#1
The code for the tea selection looks on point. Was unable to test the compass (due to no virtual compass) but all the code looks correct with the exception of the label for the first bearing. it should be changed from 
	case (theBearing >= 0 && theBearing < 23) :
 			headingText = 'North East'; 
 			break; 
to 

	case (theBearing >= 0 && theBearing < 23) :
 			headingText = 'North'; 
 			break; 

Just as a side note, perhaps try to use a different way to navigate from one window to another (besides tabs).